Chore: fjern toggle og vårrengjøring av beregning-mappen (del 1) (#5130)
### 💰 Hva skal gjøres, og hvorfor?
Favro:
https://favro.com/organization/98c34fb974ce445eac854de0/1844bbac3b6605eacc8f5543?card=NAV-24475

Fjerner toggle for å bruke ny måte å oppdatere andeler med endret
utbetaling andeler. Sletter da også masse kode som ikke blir brukt. I
tillegg er det MYE rusk i beregning-mappen, så jeg gjør store
opprydninger her. Det går hovedsakelig ut på å flytte funksjoner og
filer til andre steder de hører hjemme enn under beregning. Det er
fortsatt mye rusk å ta i, så dette blir del 1 også kommer det mer
opprydning etterhvert. End goal er at beregning-mappen kun inneholder
ting knyttet til beregning av tilkjent ytelse (kan etter opprydning også
vurdere om det er hensiktsmessig at den endrer navn til feks
tilkjentYtelse)

### 🔎️ Er det noe spesielt du ønsker tilbakemelding om?
Ja, jeg har per nå kalt flere filer -Utils når jeg har flyttet
eksisterende utils-funksjoner til nye filer. Veldig åpen for å ikke
kalle filene for Utils, men kom ikke på noe bedre i farten 😊

### ✅ Checklist
_Har du husket alle punktene i listen?_
- [x] Jeg har testet mine endringer i henhold til akseptansekriteriene
🕵️
- [ ] Jeg har config- eller sql-endringer. I så fall, husk manuell
deploy til miljø for å verifisere endringene.
- [ ] Jeg har skrevet tester. Hvis du ikke har skrevet tester, beskriv
hvorfor under 👇

_Jeg har ikke skrevet tester fordi:_
Finnnes masse tester, jeg har egentlig bare flyttet på kode og forenklet
noe logikk.

### 💬 Ønsker du en muntlig gjennomgang?
- [ ] Ja
- [x] Nei, trenger ikke det, men kan ta en gjennomgang om noen ønsker
