+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/autovedtak/småbarnstillegg/AutovedtakSmåbarnstilleggService.kt
@@ -18,13 +18,10 @@ import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingStatus
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.behandling.settpåvent.SettPåVentService
-import no.nav.familie.ba.sak.kjerne.beregning.BeregningService
-import no.nav.familie.ba.sak.kjerne.beregning.SmåbarnstilleggService
-import no.nav.familie.ba.sak.kjerne.beregning.VedtaksperiodefinnerSmåbarnstilleggFeil
-import no.nav.familie.ba.sak.kjerne.beregning.erEndringIOvergangsstønadFramITid
-import no.nav.familie.ba.sak.kjerne.beregning.finnAktuellVedtaksperiodeOgLeggTilSmåbarnstilleggbegrunnelse
-import no.nav.familie.ba.sak.kjerne.beregning.hentInnvilgedeOgReduserteAndelerSmåbarnstillegg
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakService
+import no.nav.familie.ba.sak.kjerne.grunnlag.overgangsstønad.OvergangsstønadService
+import no.nav.familie.ba.sak.kjerne.grunnlag.overgangsstønad.erEndringIOvergangsstønadFramITid
+import no.nav.familie.ba.sak.kjerne.småbarnstillegg.SmåbarnstilleggService
 import no.nav.familie.ba.sak.kjerne.steg.StegService
 import no.nav.familie.ba.sak.kjerne.steg.StegType
 import no.nav.familie.ba.sak.kjerne.vedtak.VedtakService
@@ -32,7 +29,6 @@ import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.VedtaksperiodeHentOgPe
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.VedtaksperiodeService
 import no.nav.familie.ba.sak.sikkerhet.SikkerhetContext
 import no.nav.familie.ba.sak.task.IverksettMotOppdragTask
-import no.nav.familie.ba.sak.task.OpprettTaskService
 import no.nav.familie.ba.sak.task.dto.ManuellOppgaveType
 import no.nav.familie.prosessering.internal.TaskService
 import org.slf4j.LoggerFactory
@@ -46,16 +42,15 @@ class AutovedtakSmåbarnstilleggService(
     private val vedtakService: VedtakService,
     private val behandlingService: BehandlingService,
     private val vedtaksperiodeService: VedtaksperiodeService,
-    private val småbarnstilleggService: SmåbarnstilleggService,
+    private val overgangsstønadService: OvergangsstønadService,
     private val taskService: TaskService,
-    private val beregningService: BeregningService,
     private val autovedtakService: AutovedtakService,
     private val oppgaveService: OppgaveService,
     private val vedtaksperiodeHentOgPersisterService: VedtaksperiodeHentOgPersisterService,
     private val localDateProvider: LocalDateProvider,
     private val påVentService: SettPåVentService,
     private val stegService: StegService,
-    private val opprettTaskService: OpprettTaskService,
+    private val småbarnstilleggService: SmåbarnstilleggService,
 ) : AutovedtakBehandlingService<SmåbarnstilleggData> {
     private val antallVedtakOmOvergangsstønad: Counter =
         Metrics.counter("behandling", "saksbehandling", "hendelse", "smaabarnstillegg", "antall", "aarsak", "ikke_relevant", "beskrivelse", "ikke_relevant")
@@ -88,7 +83,7 @@ class AutovedtakSmåbarnstilleggService(
 
     override fun skalAutovedtakBehandles(behandlingsdata: SmåbarnstilleggData): Boolean {
         val fagsak = fagsakService.hentNormalFagsak(aktør = behandlingsdata.aktør) ?: return false
-        val påvirkerFagsak = småbarnstilleggService.vedtakOmOvergangsstønadPåvirkerFagsak(fagsak)
+        val påvirkerFagsak = overgangsstønadService.vedtakOmOvergangsstønadPåvirkerFagsak(fagsak)
         return if (!påvirkerFagsak) {
             antallVedtakOmOvergangsstønadPåvirkerIkkeFagsak.increment()
 
@@ -108,8 +103,8 @@ class AutovedtakSmåbarnstilleggService(
             fagsakService.hentNormalFagsak(aktør)
                 ?: throw Feil(message = "Fant ikke fagsak av typen NORMAL for aktør ${aktør.aktørId}")
         val forrigeBehandling = behandlingHentOgPersisterService.hentSisteBehandlingSomErVedtatt(fagsak.id)
-        val perioderMedFullOvergangsstønadForrigeBehandling = forrigeBehandling?.let { småbarnstilleggService.hentPerioderMedFullOvergangsstønad(forrigeBehandling) } ?: emptyList()
-        val perioderMedFullOvergangsstønad = småbarnstilleggService.hentPerioderMedFullOvergangsstønad(aktør)
+        val perioderMedFullOvergangsstønadForrigeBehandling = forrigeBehandling?.let { overgangsstønadService.hentPerioderMedFullOvergangsstønad(forrigeBehandling) } ?: emptyList()
+        val perioderMedFullOvergangsstønad = overgangsstønadService.hentPerioderMedFullOvergangsstønad(aktør)
 
         val erEndringIOvergangsstønadFramITid =
             erEndringIOvergangsstønadFramITid(
@@ -174,30 +169,11 @@ class AutovedtakSmåbarnstilleggService(
     ) {
         val sistIverksatteBehandling =
             behandlingHentOgPersisterService.hentSisteBehandlingSomErIverksatt(fagsakId = behandlingEtterBehandlingsresultat.fagsak.id)
-        val forrigeSmåbarnstilleggAndeler =
-            if (sistIverksatteBehandling == null) {
-                emptyList()
-            } else {
-                beregningService
-                    .hentAndelerTilkjentYtelseMedUtbetalingerForBehandling(
-                        behandlingId = sistIverksatteBehandling.id,
-                    ).filter { it.erSmåbarnstillegg() }
-            }
-
-        val nyeSmåbarnstilleggAndeler =
-            if (sistIverksatteBehandling == null) {
-                emptyList()
-            } else {
-                beregningService
-                    .hentAndelerTilkjentYtelseMedUtbetalingerForBehandling(
-                        behandlingId = behandlingEtterBehandlingsresultat.id,
-                    ).filter { it.erSmåbarnstillegg() }
-            }
 
         val (innvilgedeMånedPerioder, reduserteMånedPerioder) =
-            hentInnvilgedeOgReduserteAndelerSmåbarnstillegg(
-                forrigeSmåbarnstilleggAndeler = forrigeSmåbarnstilleggAndeler,
-                nyeSmåbarnstilleggAndeler = nyeSmåbarnstilleggAndeler,
+            småbarnstilleggService.finnInnvilgedeOgReduserteAndelerSmåbarnstillegg(
+                behandling = behandlingEtterBehandlingsresultat,
+                sistIverksatteBehandling = sistIverksatteBehandling,
             )
 
         vedtaksperiodeHentOgPersisterService.lagre(
