+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/TilkjentYtelseGenerator.kt
@@ -1,9 +1,8 @@
 package no.nav.familie.ba.sak.kjerne.beregning
 
 import no.nav.familie.ba.sak.common.Feil
+import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ba.sak.kjerne.beregning.AndelTilkjentYtelseMedEndretUtbetalingGenerator.lagAndelerMedEndretUtbetalingAndeler
-import no.nav.familie.ba.sak.kjerne.beregning.TilkjentYtelseUtils.oppdaterTilkjentYtelseMedEndretUtbetalingAndelerGammel
-import no.nav.familie.ba.sak.kjerne.beregning.UtvidetBarnetrygdUtil.finnUtvidetVilkår
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ba.sak.kjerne.beregning.domene.EndretUtbetalingAndelMedAndelerTilkjentYtelse
@@ -11,25 +10,29 @@ import no.nav.familie.ba.sak.kjerne.beregning.domene.InternPeriodeOvergangsstøn
 import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.endretutbetaling.domene.Årsak
-import no.nav.familie.ba.sak.kjerne.fagsak.FagsakType
+import no.nav.familie.ba.sak.kjerne.grunnlag.overgangsstønad.OvergangsstønadService
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonType
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
+import org.springframework.stereotype.Component
 import java.time.LocalDate
 
-object TilkjentYtelseGenerator {
+@Component
+class TilkjentYtelseGenerator(
+    private val overgangsstønadService: OvergangsstønadService,
+    private val vilkårsvurderingService: VilkårsvurderingService,
+) {
     fun genererTilkjentYtelse(
-        vilkårsvurdering: Vilkårsvurdering,
+        behandling: Behandling,
         personopplysningGrunnlag: PersonopplysningGrunnlag,
         endretUtbetalingAndeler: List<EndretUtbetalingAndelMedAndelerTilkjentYtelse> = emptyList(),
-        fagsakType: FagsakType,
-        skalBrukeNyVersjonAvOppdaterAndelerMedEndringer: Boolean = true,
-        hentPerioderMedFullOvergangsstønad: (aktør: Aktør) -> List<InternPeriodeOvergangsstønad> = { _ -> emptyList() },
     ): TilkjentYtelse {
+        val vilkårsvurdering = vilkårsvurderingService.hentAktivForBehandlingThrows(behandlingId = behandling.id)
+
         val tilkjentYtelse =
             TilkjentYtelse(
-                behandling = vilkårsvurdering.behandling,
+                behandling = behandling,
                 opprettetDato = LocalDate.now(),
                 endretDato = LocalDate.now(),
             )
@@ -41,12 +44,12 @@ object TilkjentYtelseGenerator {
                 .beregnAndelerTilkjentYtelseForBarna(
                     personopplysningGrunnlag = personopplysningGrunnlag,
                     personResultater = vilkårsvurdering.personResultater,
-                    fagsakType = fagsakType,
+                    fagsakType = behandling.fagsak.type,
                 ).map {
                     if (it.person.type != PersonType.BARN) throw Feil("Prøver å generere ordinær andel for person av typen ${it.person.type}. Forventet ${PersonType.BARN}")
 
                     AndelTilkjentYtelse(
-                        behandlingId = vilkårsvurdering.behandling.id,
+                        behandlingId = behandling.id,
                         tilkjentYtelse = tilkjentYtelse,
                         aktør = it.person.aktør,
                         stønadFom = it.stønadFom,
@@ -60,27 +63,18 @@ object TilkjentYtelseGenerator {
                 }
 
         val barnasAndelerInkludertEtterbetaling3ÅrEller3MndEndringer =
-            if (skalBrukeNyVersjonAvOppdaterAndelerMedEndringer) {
-                lagAndelerMedEndretUtbetalingAndeler(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna.filter { it.årsak in listOf(Årsak.ETTERBETALING_3ÅR, Årsak.ETTERBETALING_3MND) },
-                    tilkjentYtelse = tilkjentYtelse,
-                )
-            } else {
-                oppdaterTilkjentYtelseMedEndretUtbetalingAndelerGammel(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna.filter { it.årsak in listOf(Årsak.ETTERBETALING_3ÅR, Årsak.ETTERBETALING_3MND) },
-                )
-            }
+            lagAndelerMedEndretUtbetalingAndeler(
+                andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
+                endretUtbetalingAndeler = endretUtbetalingAndelerBarna.filter { it.årsak in listOf(Årsak.ETTERBETALING_3ÅR, Årsak.ETTERBETALING_3MND) },
+                tilkjentYtelse = tilkjentYtelse,
+            )
 
         val andelerTilkjentYtelseUtvidetMedAlleEndringer =
             UtvidetBarnetrygdUtil.beregnTilkjentYtelseUtvidet(
-                utvidetVilkår = finnUtvidetVilkår(vilkårsvurdering),
                 tilkjentYtelse = tilkjentYtelse,
                 andelerTilkjentYtelseBarnaMedEtterbetaling3ÅrEller3MndEndringer = barnasAndelerInkludertEtterbetaling3ÅrEller3MndEndringer,
                 endretUtbetalingAndelerSøker = endretUtbetalingAndelerSøker,
                 personResultater = vilkårsvurdering.personResultater,
-                skalBrukeNyVersjonAvOppdaterAndelerMedEndringer = skalBrukeNyVersjonAvOppdaterAndelerMedEndringer,
             )
 
         val småbarnstilleggErMulig =
@@ -91,13 +85,13 @@ object TilkjentYtelseGenerator {
 
         val andelerTilkjentYtelseSmåbarnstillegg =
             if (småbarnstilleggErMulig) {
-                SmåbarnstilleggBarnetrygdGenerator(
-                    behandlingId = vilkårsvurdering.behandling.id,
+                SmåbarnstilleggGenerator(
                     tilkjentYtelse = tilkjentYtelse,
                 ).lagSmåbarnstilleggAndeler(
                     perioderMedFullOvergangsstønad =
                         hentPerioderMedFullOvergangsstønad(
-                            personopplysningGrunnlag.søker.aktør,
+                            søkerAktør = personopplysningGrunnlag.søker.aktør,
+                            behandling = behandling,
                         ),
                     utvidetAndeler = andelerTilkjentYtelseUtvidetMedAlleEndringer,
                     barnasAndeler = barnasAndelerInkludertEtterbetaling3ÅrEller3MndEndringer,
@@ -114,22 +108,27 @@ object TilkjentYtelseGenerator {
             }
 
         val andelerTilkjentYtelseBarnaMedAlleEndringer =
-            if (skalBrukeNyVersjonAvOppdaterAndelerMedEndringer) {
-                lagAndelerMedEndretUtbetalingAndeler(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna,
-                    tilkjentYtelse = tilkjentYtelse,
-                )
-            } else {
-                oppdaterTilkjentYtelseMedEndretUtbetalingAndelerGammel(
-                    andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
-                    endretUtbetalingAndeler = endretUtbetalingAndelerBarna,
-                )
-            }
+            lagAndelerMedEndretUtbetalingAndeler(
+                andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
+                endretUtbetalingAndeler = endretUtbetalingAndelerBarna,
+                tilkjentYtelse = tilkjentYtelse,
+            )
 
         tilkjentYtelse.andelerTilkjentYtelse.addAll(andelerTilkjentYtelseBarnaMedAlleEndringer.map { it.andel } + andelerTilkjentYtelseUtvidetMedAlleEndringer.map { it.andel } + andelerTilkjentYtelseSmåbarnstillegg.map { it.andel })
 
         return tilkjentYtelse
+    }
+
+    private fun hentPerioderMedFullOvergangsstønad(
+        søkerAktør: Aktør,
+        behandling: Behandling,
+    ): List<InternPeriodeOvergangsstønad> {
+        overgangsstønadService.hentOgLagrePerioderMedOvergangsstønadForBehandling(
+            søkerAktør = søkerAktør,
+            behandling = behandling,
+        )
+
+        return overgangsstønadService.hentPerioderMedFullOvergangsstønad(behandling)
     }
 
     private fun erSmåbarnstilleggMulig(
