+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/grunnlag/overgangsstønad/OvergangsstønadServiceTest.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ba.sak.kjerne.beregning
+package no.nav.familie.ba.sak.kjerne.grunnlag.overgangsstønad
 
 import io.mockk.Runs
 import io.mockk.every
@@ -17,8 +17,6 @@ import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelseRepository
 import no.nav.familie.ba.sak.kjerne.fagsak.Fagsak
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonType
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
-import no.nav.familie.ba.sak.kjerne.grunnlag.småbarnstillegg.PeriodeOvergangsstønadGrunnlag
-import no.nav.familie.ba.sak.kjerne.grunnlag.småbarnstillegg.PeriodeOvergangsstønadGrunnlagRepository
 import no.nav.familie.kontrakter.felles.ef.Datakilde
 import no.nav.familie.kontrakter.felles.ef.EksternPeriode
 import no.nav.familie.kontrakter.felles.ef.EksternePerioderResponse
@@ -29,7 +27,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 import java.time.LocalDate
 
-class SmåbarnstilleggServiceTest {
+class OvergangsstønadServiceTest {
     private val behandlingHentOgPersisterService = mockk<BehandlingHentOgPersisterService>()
     private val efSakRestClient = mockk<EfSakRestClient>()
     private val periodeOvergangsstønadGrunnlagRepository = mockk<PeriodeOvergangsstønadGrunnlagRepository>()
@@ -39,12 +37,12 @@ class SmåbarnstilleggServiceTest {
         mockk<AndelerTilkjentYtelseOgEndreteUtbetalingerService>()
     private val localDateProvider = mockk<LocalDateProvider>()
 
-    private lateinit var småbarnstilleggService: SmåbarnstilleggService
+    private lateinit var overgangsstønadService: OvergangsstønadService
 
     @BeforeEach
     fun setUp() {
-        småbarnstilleggService =
-            SmåbarnstilleggService(
+        overgangsstønadService =
+            OvergangsstønadService(
                 behandlingHentOgPersisterService = behandlingHentOgPersisterService,
                 efSakRestClient = efSakRestClient,
                 periodeOvergangsstønadGrunnlagRepository = periodeOvergangsstønadGrunnlagRepository,
@@ -103,7 +101,7 @@ class SmåbarnstilleggServiceTest {
         val slot = slot<List<PeriodeOvergangsstønadGrunnlag>>()
         every { periodeOvergangsstønadGrunnlagRepository.saveAll(capture(slot)) } returnsArgument 0
 
-        småbarnstilleggService.hentOgLagrePerioderMedOvergangsstønadForBehandling(
+        overgangsstønadService.hentOgLagrePerioderMedOvergangsstønadForBehandling(
             søkerAktør = søker.aktør,
             behandling = behandling,
         )
@@ -142,7 +140,7 @@ class SmåbarnstilleggServiceTest {
         val slot = slot<List<PeriodeOvergangsstønadGrunnlag>>()
         every { periodeOvergangsstønadGrunnlagRepository.saveAll(capture(slot)) } returnsArgument 0
 
-        småbarnstilleggService.hentOgLagrePerioderMedOvergangsstønadForBehandling(
+        overgangsstønadService.hentOgLagrePerioderMedOvergangsstønadForBehandling(
             søkerAktør = søker.aktør,
             behandling = behandling,
         )
