+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/cucumber/mock/CucumberMock.kt
@@ -36,7 +36,7 @@ import no.nav.familie.ba.sak.kjerne.behandling.settpåvent.SettPåVentService
 import no.nav.familie.ba.sak.kjerne.behandlingsresultat.BehandlingsresultatService
 import no.nav.familie.ba.sak.kjerne.behandlingsresultat.BehandlingsresultatSteg
 import no.nav.familie.ba.sak.kjerne.beregning.BeregningService
-import no.nav.familie.ba.sak.kjerne.beregning.SmåbarnstilleggService
+import no.nav.familie.ba.sak.kjerne.beregning.TilkjentYtelseGenerator
 import no.nav.familie.ba.sak.kjerne.beregning.TilkjentYtelseValideringService
 import no.nav.familie.ba.sak.kjerne.brev.mottaker.BrevmottakerService
 import no.nav.familie.ba.sak.kjerne.endretutbetaling.EndretUtbetalingAndelService
@@ -55,7 +55,9 @@ import no.nav.familie.ba.sak.kjerne.eøs.utbetaling.UtbetalingTidslinjeService
 import no.nav.familie.ba.sak.kjerne.eøs.utenlandskperiodebeløp.UtenlandskPeriodebeløpService
 import no.nav.familie.ba.sak.kjerne.eøs.valutakurs.AutomatiskOppdaterValutakursService
 import no.nav.familie.ba.sak.kjerne.eøs.valutakurs.ValutakursService
+import no.nav.familie.ba.sak.kjerne.grunnlag.overgangsstønad.OvergangsstønadService
 import no.nav.familie.ba.sak.kjerne.grunnlag.søknad.SøknadGrunnlagService
+import no.nav.familie.ba.sak.kjerne.småbarnstillegg.SmåbarnstilleggService
 import no.nav.familie.ba.sak.kjerne.steg.BeslutteVedtak
 import no.nav.familie.ba.sak.kjerne.steg.FerdigstillBehandling
 import no.nav.familie.ba.sak.kjerne.steg.IverksettMotOppdrag
@@ -74,7 +76,6 @@ import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
 import no.nav.familie.ba.sak.sikkerhet.SaksbehandlerContext
 import no.nav.familie.ba.sak.task.FerdigstillBehandlingTask
 import no.nav.familie.ba.sak.task.IverksettMotOppdragTask
-import no.nav.familie.ba.sak.task.OpprettTaskService
 import no.nav.familie.ba.sak.task.StatusFraOppdragTask
 import no.nav.familie.felles.utbetalingsgenerator.Utbetalingsgenerator
 import org.slf4j.Logger
@@ -131,7 +132,6 @@ class CucumberMock(
     val unleashNextMedContextService = mockUnleashNextMedContextService()
     val unleashService = mockUnleashService()
     val mockPåVentService = mockk<SettPåVentService>()
-    val opprettTaskService = mockk<OpprettTaskService>()
     val vurderingsstrategiForValutakurserRepository = mockVurderingsstrategiForValutakurserRepository()
     val brevmottakerService = mockk<BrevmottakerService>()
     val behandlingMigreringsinfoRepository = mockBehandlingMigreringsinfoRepository()
@@ -156,8 +156,8 @@ class CucumberMock(
             clockProvider = clockProvider,
         )
 
-    val småbarnstilleggService =
-        SmåbarnstilleggService(
+    val overgangsstønadService =
+        OvergangsstønadService(
             behandlingHentOgPersisterService = behandlingHentOgPersisterService,
             efSakRestClient = efSakRestClientMock,
             periodeOvergangsstønadGrunnlagRepository = periodeOvergangsstønadGrunnlagRepository,
@@ -189,13 +189,11 @@ class CucumberMock(
             fagsakService = fagsakService,
             behandlingHentOgPersisterService = behandlingHentOgPersisterService,
             tilkjentYtelseRepository = tilkjentYtelseRepository,
-            vilkårsvurderingRepository = vilkårsvurderingRepository,
             behandlingRepository = mockk(),
             personopplysningGrunnlagRepository = personopplysningGrunnlagRepository,
-            småbarnstilleggService = småbarnstilleggService,
             tilkjentYtelseEndretAbonnenter = listOf(tilpassDifferanseberegningEtterTilkjentYtelseService),
             andelerTilkjentYtelseOgEndreteUtbetalingerService = andelerTilkjentYtelseOgEndreteUtbetalingerService,
-            unleashNextMedContextService = unleashNextMedContextService,
+            tilkjentYtelseGenerator = TilkjentYtelseGenerator(overgangsstønadService, vilkårsvurderingService),
         )
 
     val utbetalingTidslinjeService = UtbetalingTidslinjeService(beregningService)
@@ -243,7 +241,7 @@ class CucumberMock(
             feilutbetaltValutaRepository = mockk(),
             behandlingHentOgPersisterService = behandlingHentOgPersisterService,
             vilkårsvurderingService = vilkårsvurderingService,
-            småbarnstilleggService = småbarnstilleggService,
+            overgangsstønadService = overgangsstønadService,
             refusjonEøsRepository = mockk(),
             integrasjonClient = mockk(),
             valutakursRepository = valutakursRepository,
@@ -368,6 +366,8 @@ class CucumberMock(
             localDateProvider = mockedDateProvider,
         )
 
+    val småbarnstilleggService = SmåbarnstilleggService(beregningService)
+
     val behandlingsresultatSteg =
         BehandlingsresultatSteg(
             behandlingHentOgPersisterService = behandlingHentOgPersisterService,
@@ -385,6 +385,7 @@ class CucumberMock(
             valutakursRepository = valutakursRepository,
             localDateProvider = mockedDateProvider,
             kompetanseRepository = kompetanseRepository,
+            småbarnstilleggService = småbarnstilleggService,
         )
 
     val saksbehandlerContext = SaksbehandlerContext("", mockk(), mockUnleashNextMedContextService())
@@ -605,16 +606,15 @@ class CucumberMock(
             vedtakService = vedtakService,
             behandlingService = behandlingService,
             vedtaksperiodeService = vedtaksperiodeService,
-            småbarnstilleggService = småbarnstilleggService,
+            overgangsstønadService = overgangsstønadService,
             taskService = taskService,
-            beregningService = beregningService,
             autovedtakService = autovedtakService,
             oppgaveService = oppgaveService,
             vedtaksperiodeHentOgPersisterService = vedtaksperiodeHentOgPersisterService,
             localDateProvider = mockedDateProvider,
             påVentService = mockPåVentService,
-            opprettTaskService = opprettTaskService,
             stegService = stegService,
+            småbarnstilleggService = småbarnstilleggService,
         )
     val iverksettMotOppdragTask = IverksettMotOppdragTask(stegService, behandlingHentOgPersisterService, taskRepository)
     val ferdigstillBehandlingTask = FerdigstillBehandlingTask(stegService = stegService, behandlingHentOgPersisterService = behandlingHentOgPersisterService)
